Move server-only functions out of uyuni.common
We added a dependency onto spacewalk.common when we added the
cfg_component context manager. This context manager does not make
sense on clients, clients can't access the server configuration.

Later on, this cfg_component context manager was used by
uyuni.common.fileutils. With this commit, fileutils is now split into two
parts:
  1. uyuni.common.fileutils: functions that can be used anywhere
  2. spacewalk.common.fileutils: functions that use cfg_component and
     can only be used on the server

To make consumption easy, spacewalk.common.fileutils re-exports
uyuni.common.fileutils. Users only need to import
spacewalk.common.fileutils and get everything in both
uyuni.common.fileutils and the server-only functions.

Fixes bsc#1222325
