+++ python/spacewalk/common/repo.py
@@ -1,6 +1,7 @@
 """
 Repository tools
 """
+
 # coding: utf-8
 import typing
 import http
