+++ python/spacewalk/common/rhnConfig.py
@@ -20,6 +20,8 @@ import glob
 import stat
 import re
 
+from contextlib import contextmanager
+
 from rhn.UserDictCase import UserDictCase
 from uyuni.common.usix import raise_with_tb
 
@@ -43,7 +45,6 @@ def warn(*args):
 
 
 class ConfigParserError(Exception):
-
     """
     Exception class we're using to expose fatal errors
     """
@@ -55,7 +56,6 @@ class ConfigParserError(Exception):
 #       the directory and form a _complete_ mapping structure.
 #       Or, if that is too difficult, take in a list of components...
 class RHNOptions:
-
     """Main options class
     The basic idea is to share the important pieces of information - the
     component and the configuration tree - across all instances of this
@@ -612,6 +612,33 @@ def runTest():
     print("=============== dump of all relevant dictionaries =================")
     test_cfg.showall()
     print("===================================================================")
+
+
+@contextmanager
+def cfg_component(component, root=None, filename=None):
+    """Context manager for rhnConfig.
+
+    :param comp: The configuration component to use in this context
+    :param root: Root directory location of configuration files, optional
+    :param filename: Configuration file, optional
+
+    There is a common pattern when using rhnConfig that consists of the following steps:
+    1. save current component: old = CFG.getComponent()
+    2. set CFG to another component: initCFG('my_component')
+    3. Read / Set configuration values
+    4. set CFG back to the previous component
+
+    This pattern can now be expressed using the ``with`` statement:
+
+    with cfg_component('my_component') as CFG:
+        print(CFG.my_value)
+    """
+    previous = CFG.getComponent()
+    initCFG(component=component, root=root, filename=filename)
+    try:
+        yield CFG
+    finally:
+        initCFG(previous)
 
 
 # ------------------------------------------------------------------------------
