+++ python/spacewalk/common/rhnMail.py
@@ -21,8 +21,7 @@ import os
 import smtplib
 
 from rhn.connections import idn_puny_to_unicode
-from uyuni.common.context_managers import cfg_component
-from spacewalk.common.rhnConfig import PRODUCT_NAME
+from spacewalk.common.rhnConfig import PRODUCT_NAME, cfg_component
 from typing import NamedTuple, Union
 
 with cfg_component(component=None) as CFG:
