+++ python/uyuni/common/context_managers.py
@@ -1,30 +1,8 @@
 """Collection of context managers for Uyuni."""
-from contextlib import contextmanager
-from spacewalk.common.rhnConfig import CFG, initCFG
 
-
-@contextmanager
-def cfg_component(component, root=None, filename=None):
-    """Context manager for rhnConfig.
-
-    :param comp: The configuration component to use in this context
-    :param root: Root directory location of configuration files, optional
-    :param filename: Configuration file, optional
-
-    There is a common pattern when using rhnConfig that consists of the following steps:
-    1. save current component: old = CFG.getComponent()
-    2. set CFG to another component: initCFG('my_component')
-    3. Read / Set configuration values
-    4. set CFG back to the previous component
-
-    This pattern can now be expressed using the ``with`` statement:
-
-    with cfg_component('my_component') as CFG:
-        print(CFG.my_value)
-    """
-    previous = CFG.getComponent()
-    initCFG(component=component, root=root, filename=filename)
-    try:
-        yield CFG
-    finally:
-        initCFG(previous)
+# This module always had a dependency on spacewalk.common.rhnConfig, which is wrong.
+# Everything in uyuni.common should work on clients as well. When we add other context
+# managers (that are available on clients), this import should be guarded.
+# Re-export to allow users to keep their import on uyuni.common.context_managers
+# pylint: disable-next=unused-import
+from spacewalk.common.rhnConfig import cfg_component
